+++ trunk/libaff/tools/affstats.cpp
@@ -7,7 +7,7 @@
  */
 
 /*
- * Copyright (c) 2005
+ * Copyright (c) 2005-2006
  *	Simson L. Garfinkel and Basis Technology, Inc. 
  *      All rights reserved.
  *
@@ -22,10 +22,7 @@
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 3. All advertising materials mentioning features or use of this software
- *    must display the following acknowledgement:
- *    This product includes software developed by Simson L. Garfinkel
- *    and Basis Technology Corp.
+ * 3. [omitted]
  * 4. Neither the name of Simson Garfinkel, Basis Technology, or other
  *    contributors to this program may be used to endorse or promote
  *    products derived from this software without specific prior written
@@ -124,7 +121,7 @@ void afstats(const char *fname)
 
     printf("%s\t",fname);
 
-    long unsigned int segsize=0;
+    uint32_t segsize=0;
 
     int64_t imagesize=0;
     int64_t blanksectors=0;
